purgue facturas

+++ src/main/java/net/ausiasmarch/wildcart/service/FacturaService.java
@@ -198,7 +198,7 @@ public class FacturaService {
         return oFacturaRepository.getTotalFactura(id_factura);
     }
 
-    public Long purge() {
+    public int purge() {
         //delete facturas without compras (maintenance)
         oAuthService.OnlyAdmins();
         return oFacturaRepository.purgeFacturas();
